+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Rework ical schedule export to include final events; also improve changing shifts sync
 
+### Fixed
+
+- Fix issue when creating web overrides for TF schedules using a non-UTC timezone
+
 ## v1.2.12 (2023-04-18)
 
 ### Changed